+++ packages/riot-material/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [2.0.10] - 2020-11-30
-- fix "package.json" entry point
+## [2.0.11] - 2020-12-01
+- update sub-packages
+- change rollup.config.js
 
 ## [next major] - Unreleased
 - redesign color system
@@ -20,6 +21,9 @@ All notable changes to this project will be documented in this file.
 - implement `pointerController.unbind(id: number)` function to remove the listener with given `id`
 - filterable `rm-select` does not reset value when typed
 - tabbed pages does not unmount page during motion
+
+## [2.0.10] - 2020-11-30
+- fix "package.json" entry point
 
 ## [2.0.9] - 2020-11-30
 - update all sub-packages